Fixes for the psd importer and for raising and lowering objects
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@1569 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/mspinbox.cpp
@@ -69,7 +69,7 @@ void MSpinBox::setParameters( int s )
 	if (s>=0 && s <=4)
 	{
 		Width=s;
-		Decimals=pow(10, s);
+		Decimals=static_cast<int>(pow(10, s));
 	}
 	else
 	{
